AssignmentCompiler
